# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Disaster Response Orchestrator API
  description: |
    Comprehensive REST API for disaster management, real-time alert processing, 
    AI-powered response planning, and multi-modal search capabilities using TiDB Serverless vector search.
  version: 1.0.0
  contact:
    name: AI Disaster Response Team
    email: support@disaster-orchestrator.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.disaster-orchestrator.com/api
    description: Production server
security:
  - ApiKeyAuth: []
paths:
...
  /search/hybrid:
    post:
      summary: Hybrid search
      description: Search using combined vector similarity and full-text relevance
      tags:
        - Search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HybridSearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /search/vector:
    post:
      summary: Vector similarity search
      description: Search using semantic vector similarity
      tags:
        - Search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VectorSearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  /search/fulltext:
    post:
      summary: Full-text search
      description: Search using MySQL full-text search with relevance scoring
      tags:
        - Search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FullTextSearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key authentication (Bearer token)
  schemas:
...
    HybridSearchRequest:
      type: object
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 500
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
        threshold:
          type: number
          minimum: 0
          maximum: 1
          default: 0.7
        filters:
          type: object
          properties:
            category:
              type: string
            type:
              type: string
            location:
              type: string
        weights:
          type: object
          properties:
            vector:
              type: number
              minimum: 0
              maximum: 1
              default: 0.4
            text:
              type: number
              minimum: 0
              maximum: 1
              default: 0.6
      required:
        - query
    VectorSearchRequest:
      type: object
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 500
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
        threshold:
          type: number
          minimum: 0
          maximum: 1
          default: 0.8
        filters:
          type: object
      required:
        - query
    FullTextSearchRequest:
      type: object
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 500
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
        filters:
          type: object
      required:
        - query
...
    SearchResult:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        category:
          type: string
        similarity:
          type: number
          description: Vector similarity score
        text_score:
          type: number
          description: Full-text relevance score
        combined_score:
          type: number
          description: Weighted combined score
        metadata:
          type: object
...
    SearchResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        query_info:
          type: object
          properties:
            processed_query:
              type: string
            embedding_generated:
              type: boolean
            search_type:
              type: string
            total_results:
              type: integer
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request - validation errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Validation failed"
            code: "VALIDATION_ERROR"
            details:
              query: "Query must be between 10 and 500 characters"
            timestamp: "2025-09-15T12:00:00Z"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Internal server error"
            code: "INTERNAL_ERROR"
            timestamp: "2025-09-15T12:00:00Z"
...